+++ src/commands/generate.rs
@@ -1,21 +1,25 @@
 use anyhow::Context;
-use clap::{value_t, ArgMatches};
+use clap::ArgMatches;
 use rand::{rngs::SmallRng, SeedableRng};
 use tracing::info;
 
 use crate::{generator::Builder, Generator, PairWriter};
 
-pub fn generate(matches: &ArgMatches<'_>) -> anyhow::Result<()> {
+pub fn generate(matches: &ArgMatches) -> anyhow::Result<()> {
     let r1_dst = matches.value_of("r1-dst").unwrap();
     let r2_dst = matches.value_of("r2-dst").unwrap();
 
-    let record_count = value_t!(matches, "record-count", u64).unwrap_or_else(|e| e.exit());
-    let read_length = value_t!(matches, "read-length", usize).unwrap_or_else(|e| e.exit());
+    let record_count = matches
+        .value_of_t("record-count")
+        .unwrap_or_else(|e| e.exit());
+    let read_length = matches
+        .value_of_t("read-length")
+        .unwrap_or_else(|e| e.exit());
 
     info!("fq-generate start");
 
     let builder = if matches.is_present("seed") {
-        let seed = value_t!(matches, "seed", u64).unwrap_or_else(|e| e.exit());
+        let seed = matches.value_of_t("seed").unwrap_or_else(|e| e.exit());
         let rng = SmallRng::seed_from_u64(seed);
         Builder::from_rng(rng)
     } else {