+++ CHANGELOG.md
@@ -30,7 +30,7 @@
     This allows subcommands to show the global version, e.g., `fq lint
     --version`.
 
-[#20]: https://github.com/stjude/fqlib/issues/20
+[#20]: https://github.com/stjude-rust-labs/fq/issues/20
 
 ## 0.7.0 - 2021-05-07